Added CSS variables for setting the option text color normal and on hover/focus. Adjusted the padding on the button.
Changed the multi select constructor so that it can accept null for the label. If null is passed in then a label is created.

Added an onChangeCallback function that can be set through the constructor or through a public function registerOnChangeCallback.
This function is passed the multiSelectID and a list of selected values.